+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.7 (unreleased)
 
+- Add import to `catalystcenter_assign_templates_to_tag` resource
 - Add `catalystcenter_assign_devices_to_tag` resource and data source
 
 ## 0.2.6